% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_smlc.R
\name{adjust_Xnew}
\alias{adjust_Xnew}
\title{adjust Xnew by discarding columns not in Xold_colnames
and by adding 0-valued columns that are in Xold_colnames
but not in Xnew}
\usage{
adjust_Xnew(Xnew, Xold_colnames)
}
\description{
adjust Xnew by discarding columns not in Xold_colnames
and by adding 0-valued columns that are in Xold_colnames
but not in Xnew
}
