% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_calc_minfo.R
\name{calc_minfo}
\alias{calc_minfo}
\title{calculates normalized mutual information given P(variant | cancer)
and p(cancer)}
\usage{
calc_minfo(prob_v_given_c, prob_c, normalize = TRUE)
}
\description{
calculates normalized mutual information given P(variant | cancer)
and p(cancer)
}
