% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nn_helper.R
\name{create_model}
\alias{create_model}
\title{Initialize a fully-connected multi-class neural network}
\usage{
create_model(
  learning_rate = 0.001,
  weight_decay = 0,
  dropout = 0,
  num_dense_layers = 1,
  num_dense_nodes = 5,
  activation = "relu",
  seed = NULL,
  input_shape,
  num_classes
)
}
\arguments{
\item{learning_rate}{Learning rate for Adam optimizer}

\item{weight_decay}{L2 regulariation paramter}

\item{dropout}{Dropout rate}

\item{num_dense_layers}{Number of hidden layers}

\item{num_dense_nodes}{Number of nodes per hidden layer (all hidden layers have the same size)}

\item{activation}{Name of activation function ("relu", "softmax", "softplus", "elu", "tanh", "sigmoid", etc)}

\item{seed}{Seed for weight initialization and dropout}

\item{input_shape}{Number of features}

\item{num_classes}{Number of classes}
}
\value{
A compiled Keras model
}
\description{
Initialize a fully-connected multi-class neural network
}
\examples{
model <- create_model(learning_rate = 0.001, weight_decay = 0.5, dropout = 0.2, num_dense_layers = 2, num_dense_nodes = 20, activation = "relu", seed = 1, input_shape = 1324, num_classes = 8)
summary(model)
}
\author{
Zoe Guan
}
