% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_rows_cols.R
\name{divide_rows}
\alias{divide_rows}
\alias{divide_cols}
\title{Convenience functions for dividing rows and columns of a matrix
by a given vector}
\usage{
divide_rows(mat, vec)

divide_cols(mat, vec)
}
\arguments{
\item{mat}{A matrix. Can be a sparseMatrix from {Matrix}}

\item{vec}{}
}
\description{
Convenience functions for dividing rows and columns of a matrix
by a given vector
}
\details{
\code{divide_rows} (\code{divide_cols}) divide the i-th row
(i-th column) of \code{mat} by the i-th entry of \code{vec}.
Formally, \code{divide_rows(mat, vec)} is equivalent to
sweep(mat, 1, vec "/"), but is usually more efficient, and
plays nicer with {magrittr} pipe \code{\%>\%}.
}
