% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_multiclass.R
\name{do_cv_multiclass_allmodels}
\alias{do_cv_multiclass_allmodels}
\title{Cross validation smlc experiments with n_folds
INPUT:
maf_dt_in: full mutation annotation file (maf)
all_covmats: list of covariance matrices
(XU combined) for patients
with colum names
g = gene, v = variants, c = cancer type,
and p = patient (tumor) id. Will be used to compute the
design matrix of recorded variants in each cross validation fold.
XU_g, XU_ms: desgin-meta-design matrix with genes and mutation
signatures respectively.}
\usage{
do_cv_multiclass_allmodels(
  maf_dt_in,
  all_covmats,
  covmat_append_unsc_vardesign = rep(FALSE, length(all_covmats)),
  covmat_append_sc_vardesign = rep(FALSE, length(all_covmats)),
  covmat_append_sqrt_TMB = rep(FALSE, length(all_covmats)),
  Y,
  sqrt_TMB,
  g_list,
  v_rank_thresh_full = 800,
  n_fold = 5,
  alpha = 1,
  grouped = FALSE,
  mi_normalized = TRUE,
  recorded_var_classifier = FALSE,
  scale_recorded_var_classifer = FALSE,
  ...
)
}
\description{
Cross validation smlc experiments with n_folds
INPUT:
maf_dt_in: full mutation annotation file (maf)
all_covmats: list of covariance matrices
(XU combined) for patients
with colum names
g = gene, v = variants, c = cancer type,
and p = patient (tumor) id. Will be used to compute the
design matrix of recorded variants in each cross validation fold.
XU_g, XU_ms: desgin-meta-design matrix with genes and mutation
signatures respectively.
}
