% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_mat.R
\name{extract_cancer}
\alias{extract_cancer}
\title{Extract cancer labels}
\usage{
extract_cancer(maf)
}
\description{
Extract cancer labels
}
\details{
INPUT:
maf: mutation annotation file (maf) with colum names
g = gene, v = variants, c = cancer type,
and p = patient (tumor) id.

OUTPUT:
named (by patient_id) vector of cancer labels
}
