% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_design_mdesign_mcat.R
\name{extract_design_mdesign_mcat}
\alias{extract_design_mdesign_mcat}
\title{Extract Design-Metadesign (XU) matrix for a categorical
meta-feature from a maf file}
\usage{
extract_design_mdesign_mcat(
  maf,
  variant_col = "variant",
  mfeat_col = "gene",
  sample_id_col = "sample",
  mfeat_subset = NULL,
  ...
)
}
\arguments{
\item{maf}{mutation annotation file --
a data frame-like object with at least three columns containing
variant labels, sample IDs and (categorical) meta-feature labels.
NOTE: uniqueness of rows of maf is assumed.}

\item{variant_col}{name of the column in \code{maf} containing variant labels.}

\item{mfeat_col}{name of the column in \code{maf} containing categorical
meta-feature labels.}

\item{sample_id_col}{name of the column in \code{maf} containing tumor sample IDs.}

\item{...}{Unused.}
}
\description{
Extract Design-Metadesign (XU) matrix for a categorical
meta-feature from a maf file
}
