% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_design_mdesign_mcts.R
\name{extract_design_mdesign_mcts}
\alias{extract_design_mdesign_mcts}
\title{Extract Design-Metadesign (XU) matrix for a set of continuous
meta-features stored in a maf file}
\usage{
extract_design_mdesign_mcts(
  maf,
  variant_col = "variant",
  mfeat_cols = "Histone_Marks",
  sample_id_col = "sample",
  ...
)
}
\arguments{
\item{maf}{mutation annotation file --
a data frame-like object containing columns for
variant labels, sample IDs and continuous meta-feature variables.
NOTE: uniqueness of rows of maf is assumed}

\item{variant_col}{name of the column in \code{maf} containing variant labels.}

\item{mfeat_cols}{names of columns in \code{maf} containing continuous
meta-feature variables. Can be a vector.}

\item{sample_id_col}{name of the column in \code{maf} containing tumor sample IDs.}

\item{...}{Unused.}
}
\description{
Extract Design-Metadesign (XU) matrix for a set of continuous
meta-features stored in a maf file
}
