% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_smlc.R
\name{extract_glmnet}
\alias{extract_glmnet}
\title{Extract the cv.glmnet object from the output of fit_mlogit}
\usage{
extract_glmnet(fit, ...)
}
\arguments{
\item{fit}{fitted hidden genome mlogit classifier, an output of fit_smlc.}
}
\description{
Extract the cv.glmnet object from the output of fit_mlogit
}
