% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_TMB.R
\name{extract_tmb}
\alias{extract_tmb}
\title{Extract total mutation burden (TMB) of tumors in a maf file}
\usage{
extract_tmb(maf, variant_col = "variant", sample_id_col = "sample", ...)
}
\arguments{
\item{maf}{mutation annotation file --
a data frame-like object with at least two columns containing
variant labels and sample IDs. NOTE: uniqueness of rows of maf is assumed.}

\item{variant_col}{name of the column in \code{maf} containing variant labels.}

\item{sample_id_col}{name of the column in \code{maf} containing tumor sample IDs.}

\item{...}{Unused.}
}
\value{
Returns a numeric vector providing the total
number of mutations observed in each tumor in maf, named
by the sample ids in maf.
}
\description{
Extract total mutation burden (TMB) of tumors in a maf file
}
