% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_rfc.R
\name{fit_rfc}
\alias{fit_rfc}
\title{Hidden genome random forest classifier (rfc)}
\usage{
fit_rfc(X, Y, backend = "ranger", ...)
}
\arguments{
\item{X}{data design matrix with observations across rows and predictors across
columns. For a typical hidden genome classifier each row represents a tumor and
the columns represent (possibly normalized by some functions of
the total mutation burden in tumors) binary 1-0 presence/absence indicators
of raw variants, counts of mutations at specific genes and counts of mutations
corresponding to specific mutation signatures etc.}

\item{Y}{character vector or factor denoting the cancer type of tumors whose
mutation profiles are listed across the rows of \code{X}.}

\item{backend}{Which backend to use? Available options are
"ranger" and "randomForest" corresponding to the respective R packages.
NOTE: randomForest does not support sparseMatrix, and the predictor matrix
is coerced into an ordinary matrix. This means using randomForest will likely
be more memory intensive and hence slower than ranger.}

\item{...}{additional arguments passed to ranger or randomForest (depending
on backend).}
}
\description{
Hidden genome random forest classifier (rfc)
}
\details{
Light wrapper around randomForest or ranger to use in hidden
genome classification
}
