% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sml.R
\name{fit_smlc}
\alias{fit_smlc}
\title{Hidden genome Bayesian hierarchical classifier}
\usage{
fit_smlc(X, Y, grouped = TRUE, alpha = 1, no_penalty_vars = NULL, ...)
}
\description{
Light wrapper around cv.glmnet with  family = "multinomia",
and type.mutlinomial = "grouped" if grouped = TRUE.
Returns the glmnet fitted object together with X, Y and estiamted
intercept alpha and regression coefficients beta
}
