% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nn_helper.R
\name{fitness}
\alias{fitness}
\title{Fit and validate neural network}
\usage{
fitness(
  X_train_norm,
  Y_train_onehot,
  X_val_norm,
  Y_val_onehot,
  learning_rate,
  weight_decay,
  dropout,
  num_dense_layers,
  num_dense_nodes,
  activation,
  epochs = 50,
  batch_size = 32,
  verbose = 0,
  seed = NULL
)
}
\arguments{
\item{X_train_norm}{Normalized input matrix for training set}

\item{Y_train_onehot}{One-hot representation of response for training set}

\item{X_val_norm}{Normalized input matrix for validation set}

\item{Y_val_onehot}{One-hot representation of response for validation set}

\item{epochs}{Number of training epochs}

\item{batch_size}{Batch size}

\item{verbose}{Keras verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch)}
}
\value{
Validation accuracy
}
\description{
Fit and validate neural network
}
