% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_prediction_effect.R
\name{indiv_predict_effect}
\alias{indiv_predict_effect}
\title{Probability distance based effects of predictors in a fitted model
in each \emph{individual prediction}}
\usage{
indiv_predict_effect(fit, Xnew, dist = "jsdist", ...)
}
\arguments{
\item{dist}{Probability distance measure to use. Defaults to "jsdist"
which is the Jensen-Shannon distance}

\item{...}{additional arguments passed to \code{cv.glmnet}.}
}
\description{
Probability distance based effects of predictors in a fitted model
in each \emph{individual prediction}
}
