% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_mat.R
\name{make_design}
\alias{make_design}
\title{Find design (or design-meta-design) matrix for
variants and genes}
\usage{
make_design(
  maf,
  column = "v",
  sampleid_col = "p",
  col_select = "all",
  calc_nontarget_p = TRUE,
  target_g = "all"
)
}
\description{
Find design (or design-meta-design) matrix for
variants and genes
}
\details{
INPUT:
maf: mutation annotation file (maf) with colum names
g = gene, v = variants, c = cancer type,
and p = patient (tumor) id.
column: name of the column that corresponds
to variants. Either "v" (for variants, if design matrix) or
"g" for genes (if design-meta-design)
col_select: either "all", which means all columns of the design
matrix will be retained, based on all variants/genes observed in maf,
or a character vector of variants/genes that are to be retained.

OUTPUT:
an N_patients x N_variants (or N_genes) (sparse) design
(or design-meta-design) matrix with appropriate row and
column names.
}
