% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_nnc.R
\name{new_nn}
\alias{new_nn}
\title{Generate an object from the class "nn"}
\usage{
new_nn(map_df, model, model_raw, ind_val, tuning_results, preproc)
}
\description{
See description of output of fit_nn() for details
}
