% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_or_mean_1sd.R
\name{odds_ratio_smlc}
\alias{odds_ratio_smlc}
\title{Calculate odds ratios from a multinomial logistic
hidden genome model}
\usage{
odds_ratio_smlc(
  fit,
  type = "one-vs-rest",
  scale = TRUE,
  log = TRUE,
  predictor_subset = NULL,
  baseline_category = NULL,
  ...
)
}
\arguments{
\item{type}{either "one-vs-rest" or "one-vs-one".}

\item{baseline_category}{The response category with respect to which
odds ratios are calculated from the fitted model if \code{type = "one-vs-one"}.
If NULL (default), the first response cancer category obtained after sorting the
column labels (using \code{sort()}) is used.}
}
\description{
Calculate odds ratios from a multinomial logistic
hidden genome model
}
\details{
If  \code{type = "one-vs-rest"}, odds ratios are calculated for each predictor,
across all response categories, for one unit (one standard deviation unit,
if \code{scale = TRUE}) increase in each predictor at its mean, while keeping
all other predictors fixed at their respective means.

If \code{type = "one-vs-one"}, odds ratios relative to a baseline category
is calculated.
}
