% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_nnc.R
\name{predict_nnc}
\alias{predict_nnc}
\title{Get neural network predictions}
\usage{
predict_nnc(fit, Xnew, Ynew = NULL, ...)
}
\arguments{
\item{fit}{Fitted neural network hidden genome classifier
(output of fit_nnc())}

\item{...}{additional arguments passed to \code{cv.glmnet}.}
}
\value{
list with entries:
}
\description{
Get neural network predictions
}
