% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sml.R
\name{predict_rfc}
\alias{predict_rfc}
\title{prediction based on hidden genome random forest classifier}
\usage{
predict_rfc(Xnew, fit, Ynew = NULL, ...)
}
\description{
prediction based on hidden genome random forest classifier
}
