% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sml.R
\name{predict_smlc}
\alias{predict_smlc}
\title{Prediction based on sparse multilevel multinomial
logisitic classifier}
\usage{
predict_smlc(Xnew, fit, Ynew = NULL, return_lin_pred = FALSE, ...)
}
\description{
can handles unseen variants as predictors.
}
