% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_svmc.R
\name{predict_svmc}
\alias{predict_svmc}
\title{prediction based on hidden genome random forest classifier}
\usage{
predict_svmc(fit, Xnew, Ynew = NULL, ...)
}
\description{
prediction based on hidden genome random forest classifier
}
