% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_mat.R
\name{select_v_nmi}
\alias{select_v_nmi}
\title{compute NMIs of the variants & select variants based on NMI ranking}
\usage{
select_v_nmi(
  dat,
  nmi_rank_thresh,
  column = "v",
  cancer_col = "c",
  patient_col = "p",
  return_probs = TRUE,
  normalize_mi = TRUE
)
}
\description{
compute NMIs of the variants & select variants based on NMI ranking
}
\details{
INPUT:
g = gene, v = variants, c = cancer type,
and p = patient (tumor) id.
nmi_rank_thresh: threshold for NMI ranks,
variants with NMI rank <= nmi_rank_thresh will be kept
select v's based on mi ranking
column: name of the column in dat that corresponds to variants ("v")
return_probs: logical. return the relative and NMI values?

OUTPUT: a list with elements:
v_select: character vector of names of variants with
ranks <= nmi_rank_thresh.
probs_mi: conditional proabilities of the variants given
cancersAlso returns
}
