% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nn_helper.R
\name{tune_model}
\alias{tune_model}
\title{Tune neural network hyperparameters using Bayesian optimization}
\usage{
tune_model(
  X_train_norm,
  Y_train_onehot,
  X_val_norm,
  Y_val_onehot,
  trials = 200,
  epochs = 50,
  verbose_keras = 0,
  verbose_mbo = T,
  seed = NULL
)
}
\arguments{
\item{trials}{Number of trials}

\item{verbose_keras}{Keras verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch)}

\item{verbose_mbo}{Bayesian optimization verbosity mode (logical)}
}
\value{
See description of mbo() function from mlrMBO
}
\description{
Tune neural network hyperparameters using Bayesian optimization
}
