% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_threshold.R
\name{calc_one_v_rest_auc}
\alias{calc_one_v_rest_auc}
\title{Calculating area under Precision-Recall curve (PRC) and
Receiver-Operator characteristic curve (ROC) for all one-vs-rest
comparisons in the fitted model}
\usage{
calc_one_v_rest_auc(fit, measure = c("PRC", "ROC"), fitted_prob = NULL, ...)
}
\arguments{
\item{fit}{fitted hidden genome classifier object}

\item{measure}{prediction assessment measure. Options include "fscore",
"mcc" (Mathews Correlation Coefficient). Can be a vector.}
}
\value{
Returns a data.table with \code{length(measure) + 1} columns
("Class" and measure(s)) and n_class + 1 many rows, where n_class
denotes the number of cancer types present in the fitted model; the
final row provides the Macro (average) metrics.
}
\description{
Calculating area under Precision-Recall curve (PRC) and
Receiver-Operator characteristic curve (ROC) for all one-vs-rest
comparisons in the fitted model
}
