% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_smlc.R
\name{odds_ratio_mlogit}
\alias{odds_ratio_mlogit}
\alias{odds_ratio_smlc}
\title{Calculate odds ratios from a multinomial logistic
hidden genome model}
\usage{
odds_ratio_mlogit(
  fit,
  type = "one-vs-rest",
  scale = FALSE,
  log = TRUE,
  predictor_subset = NULL,
  baseline_category = NULL,
  ...
)

odds_ratio_smlc(
  fit,
  type = "one-vs-rest",
  scale = FALSE,
  log = TRUE,
  predictor_subset = NULL,
  baseline_category = NULL,
  ...
)
}
\arguments{
\item{fit}{fitted hidden genome mlogit classifier, an output of fit_smlc.}

\item{type}{either "one-vs-rest" or "one-vs-one".}

\item{log}{logical. Should the odds ratios be returned in log scale?
Defaults to TRUE.}

\item{predictor_subset}{Character vector listing the subset of predictors in
the fitted model \code{fit} for which odds ratios are to be computed.}

\item{baseline_category}{The response category with respect to which
odds ratios are calculated from the fitted model if \code{type = "one-vs-one"}.
If NULL (default), the first response cancer category obtained after sorting the
column labels (using \code{sort()}) is used.}
}
\value{
Returns a sparse matrix (of class dgeMatrix) with odds ratios for predictors
(along the rows) across cancer sites (along the columns).
}
\description{
Calculate odds ratios from a multinomial logistic
hidden genome model
}
\details{
If  \code{type = "one-vs-rest"}, odds ratios are calculated for each predictor,
across all response categories, for one unit (one standard deviation unit,
if \code{scale = TRUE}) increase in each predictor at its mean, while keeping
all other predictors fixed at their respective means.

If \code{type = "one-vs-one"}, odds ratios relative to a baseline category
is calculated. (Not implemented yet.)
}
