% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_threshold.R
\name{optimal_threshold}
\alias{optimal_threshold}
\title{Determine optimal one-vs-rest classification
thresholds from fitted hidden genome models using
prediction-based performance measures}
\usage{
optimal_threshold(fit, measure = "fscore", fitted_prob = NULL, ...)
}
\arguments{
\item{fit}{fitted hidden genome classifier object}

\item{measure}{prediction assessment measure. Options include "fscore",
"mcc" (Mathews Correlation Coefficient). Can be a vector.}
}
\value{
If \code{length(measure) == 1} the function returns a named vector with optimal
one-vs-rest classification thresholds for all cancer
classes in the fitted hidden genome object (fit). The optimal
values obtained at the corresponding optimal thresholds are
returned as an attribute "optimal_value".

If \code{length(measure) > 1} a named list is returned, with each
entry providing the optimal thresholds across all cancer categories
(along with the associated optimal measure values as an attribute)
corresponding to that \code{measure}.
}
\description{
Determine optimal one-vs-rest classification
thresholds from fitted hidden genome models using
prediction-based performance measures
}
