% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_rfc.R
\name{predict_rfc}
\alias{predict_rfc}
\alias{predict_rf}
\title{prediction based on hidden genome random forest classifier}
\usage{
predict_rfc(fit, Xnew, Ynew = NULL, ...)

predict_rf(fit, Xnew, Ynew = NULL, ...)
}
\arguments{
\item{fit}{Fitted random forest hidden genome classifier (output of
fit_rfc).}
}
\description{
prediction based on hidden genome random forest classifier
}
