% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_svmc.R
\name{predict_svmc}
\alias{predict_svmc}
\title{prediction based on hidden genome random forest classifier}
\usage{
predict_svmc(fit, Xnew, Ynew = NULL, ...)
}
\arguments{
\item{fit}{fitted hidden genome SVM classifier (output of
\code{fit_svmc()})}

\item{Xnew}{test data design (or meta-design) matrix (observations
across rows and variables predictors/features across columns)
for which predictions are to be made from a fitted model. For a typical hidden
genome classifier this will be a matrix whose rows correspond to the test set
tumors, and columns correspond to (normalized by some functions of
the total mutation burdens in tumors) binary 1-0 presence/absence of
raw variants, counts of mutations at specific genes and counts of mutations
corresponding to specific mutation signatures etc.}

\item{Ynew}{the actual cancer categories for the test samples.
This is not used in computation, but is return as a component in the output,
for possibly easier post-processing.}
}
\description{
prediction based on hidden genome random forest classifier
}
\seealso{
fit_svmc
}
